+++ knexfile.js
@@ -2,6 +2,13 @@
 // npx knex init (to initialise a knexfile.js)
 // run below command to generate tables in DB
 //npx knex migrate:latest
+/*
+Add below in connection object with password, database for ssl issue
+ ssl: {
+        rejectUnauthorized: false,
+        require: true
+      }
+*/
 
 export default {
   development: {
